pyproject.toml
[tool.poetry]
name = "notionbooks"
version = "0.1.0"
description = "Turn your notion database into Jupyter Notebooks"
authors = ["Antonio Feregrino <antonio.feregrino@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
notion-client = "^2.0.0"
requests = "^2.28.2"
python-slugify = "^8.0.1"
jupyter = "^1.0.0"
typer = "^0.7.0"

# Add entrypoint for the CLI
[tool.poetry.scripts]
nbooks = "notionbooks.__main__:app"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[tool.black]
line-length = 120

[tool.isort]
skip_glob = [".venv", "dist"]
# Structure
default_section = "THIRDPARTY"
known_first_party = "notionbooks"
# Import Style
line_length = 120
force_grid_wrap = false
use_parentheses = true
include_trailing_comma = true
combine_as_imports = true
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
